refactor: complete WordPress decoupling — remaining security, headers, SEO
Security classes (6 more) — inject RequestContextInterface:
- DisableUserEnumeration: replace $_GET with requestContext->query()
- RestApiSecurity: replace $_SERVER['REQUEST_URI'] with requestContext
- PasswordPolicy: replace $_POST['pass1'] with requestContext->post()
- TwoFactorAuthentication: replace $_POST['backto_2fa_code']
- AuditLogAdminPage: replace 7 superglobal accesses ($_GET/$_POST/$_REQUEST)
- DatabaseHardening: replace $_SERVER['SCRIPT_FILENAME']

ResponseEmitter abstraction (header/exit decoupling):
- Create ResponseEmitterInterface contract + NativeResponseEmitter
- CorsManager: inject RequestContext + ResponseEmitter, remove all header()/exit
- HttpHeadersHardening: inject ResponseEmitter, remove all header() calls

SEO — BreadcrumbSchemaGenerator to Infrastructure:
- Extract BreadcrumbSchemaGeneratorInterface contract
- Move implementation to Seo/Infrastructure/WordPressBreadcrumbSchemaGenerator
- Old class kept as deprecated shim
- RegisterDefaultSchemas now depends on interface, not concrete class
